requirements.txt
pytest==5.1.2
Flask==1.1.1
pubnub==4.1.6
python-dotenv==0.13.0